Move code from upstream to insights/contrib (#860)
* Move code from upstream to insights/contrib

* Fix issue in example
